pyproject.toml
[tool.poetry]
name = "desy-bluesky"
version = "0.1.0"
description = ""
authors = ["Devin Burke"]
license = "GNU GPL 3"
packages = [
    { include = "desy_bluesky" }
]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "*"
bluesky = "*"
databroker = "*"
matplotlib = "*"
nxarray = "*"
ipython = "*"
pytest = "*"
PyYAML = "*"
pyepics = "*"
pytango = "*"
aioca = "*"
ophyd = "*"
ophyd-async = { git = "https://gitlab.desy.de/fs-ec/desy-bluesky/ophyd-async", branch = "tango_support" }

[tool.poetry.dev-dependencies]
pre-commit = "*"
poetry = "*"
ruff = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
